Updated battlesnake for 2018
Updated battlesnake for 2018. Includes a sample snake in python
