--- battlesnake-python/Procfile
@@ -1,1 +1,0 @@
-web: gunicorn app.main:application --worker-class gevent
