--- battlesnake-python/app.json
@@ -1,7 +1,0 @@
-{
-  "name": "BattleSnake Python",
-  "description": "A simple BattleSnake AI written in Python.",
-  "repository": "https://github.com/sendwithus/battlesnake-python",
-  "website": "http://battlesnake.io",
-  "keywords": ["battlesnake","sendwithus"]
-}
