--- battlesnake-python/cloud9/run.sh
@@ -1,3 +1,0 @@
-cd ..
-echo -e "\n===== https://$C9_HOSTNAME =====\n"
-gunicorn app.main:application --access-logfile - --worker-class gevent --bind "$IP:$PORT"
