--- battlesnake-python/requirements.txt
@@ -1,4 +1,0 @@
-bottle==0.12.9
-gevent==1.0.2
-greenlet==0.4.9
-gunicorn==19.4.5
